anom_server/anom_server/urls.py
"""anom_server URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/1.8/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
Including another URLconf
    1. Add an import:  from blog import urls as blog_urls
    2. Add a URL to urlpatterns:  url(r'^blog/', include(blog_urls))
"""
from django.conf.urls import include, url
from django.contrib import admin

from django.contrib.auth.models import User
from rest_framework import serializers
from rest_framework_jwt.authentication import JSONWebTokenAuthentication

from rest_framework.response import Response
from rest_framework.views import APIView


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'email', 'is_staff')


class UserProfileView(APIView):
    authentication_classes = (JSONWebTokenAuthentication,)

    def get(self, request, format=None):
        """
        Return a current_user profile
        """
        return Response(UserSerializer(request.user).data)


urlpatterns = [
    url(r'^profiles/', include('profiles.urls')),
    url(r'^api-user-me/', UserProfileView.as_view()),
    url(r'^admin/', include(admin.site.urls)),
    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    url(r'^api-token-auth/', 'rest_framework_jwt.views.obtain_jwt_token')
]
