anom_server/requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-filter
MySQL-python
django-cors-headers
